Refactor schema validations
:Release Notes:
Implemented WRO-10835
[UWB] API code refactoring

:Detailed Notes:
Refactored schema validations.

:Testing Performed:
tested locally

:QA Notes:
 NA

:Issues Addressed:
[WRO-10550] - Refactored schema validations.

Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.uwb/+/334949
Tested-by: <muralidhar.n@lge.com>
(cherry picked from commit 7679ab5ea1b600f1fc4650862d97882f3d481e50)
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.uwb/+/335865

+++ src/UwbRangingInfo.cpp
@@ -10,7 +10,7 @@ UwbRangingInfo::UwbRangingInfo() : m_sessionId(0),
     mAltitudeAngleMeasure = std::make_unique<AngleMeasure>(0.0, 0.0, 0.0);
 }
 
-uint16_t UwbRangingInfo::getSessionId() const{
+uint8_t UwbRangingInfo::getSessionId() const{
     return m_sessionId;
 }
 
@@ -42,7 +42,7 @@ std::unique_ptr<AngleMeasure>& UwbRangingInfo::getAltitudeAngleMeasure(){
     return mAltitudeAngleMeasure;
 }
 
-void UwbRangingInfo::setSessionId(uint16_t sessionId) {
+void UwbRangingInfo::setSessionId(uint8_t sessionId) {
     m_sessionId = sessionId;
 }
